edn1_2_dataviz/requirements.txt
duckdb>=1.1.0
pyarrow>=14.0.0
pandas>=2.1.0
streamlit>=1.37.0
pyyaml>=6.0.0
altair>=5.0.0
pytest>=7.4.0

